Made it good. No warnings :)
