+++ src/components/App.js
@@ -8,15 +8,11 @@ import {
   About,
   Contact,
   Ticket,
-  Posts,
-  Post,
-  Test,
   Login,
   Register,
   Profile,
   CreateTicket,
-  TicketTable,
-  DisplayTicket
+  TicketTable
 } from "../components";
 
 const App = () => {
@@ -31,15 +27,10 @@ const App = () => {
         <Route path="/about" element={<About />} />
         <Route path="/contact" element={<Contact />} />
         <Route path="/table" element={<TicketTable />} />
-        <Route path="/ticket" element={<Ticket />}>
-          <Route path="" element={<Posts />} />
-          <Route path=":postSlug" element={<Post />} />
+        <Route path="/ticket">
+          <Route path=":ticketSlug" element={<Ticket />} />
         </Route>
-        <Route path="/table" element={<TicketTable />} />
         <Route path="/createticket" element={<CreateTicket />} />
-        <Route path="/test" element={<Test />} />
-        <Route path="tickets" element={<TicketTable />} />
-        <Route path="/DisplayTicket" element={<DisplayTicket />} />
         <Route path="/login" element={<Login setId={setId} />} />
         <Route path="/register" element={<Register setId={setId} />} />
         <Route path="/profile" element={<Profile id={id} setId={setId} updateState={updateState} />} />
