+++ src/components/index.js
@@ -3,15 +3,11 @@ export { default as Footer } from "./Footer";
 export { default as Home } from "./Home";
 export { default as About } from "./About";
 export { default as Contact } from "./Contact";
-export { default as Ticket } from "./blog/Ticket";
 export { default as CreateTicket } from "./CreateTicket";
 export { default as Login } from "./Login"
 export { default as Register } from "./Register"
 export { default as Profile } from "./Profile"
-export { default as Posts } from "./blog/Posts";
-export { default as Post } from "./blog/Post";
-export { default as Test } from "./Test";
+export { default as Ticket } from "./Ticket";
 export { default as App } from "./App";
 export { default as TicketTable } from "./TicketTable";
-export { default as DisplayTicket } from "./DisplayTicket";
